+++ subsurface/io/topography/__init__.py
@@ -3,7 +3,7 @@ import ezdxf
 import numpy as np
 from scipy.spatial.qhull import Delaunay
 
-from subsurface import UnstructuredData, StructuredData
+from subsurface.structs.base_structures import UnstructuredData, StructuredData
 from subsurface.io.mesh.surface_reader import read_2d_mesh
 from subsurface.utils import get_extension
