+++ tests/test_io/test_welly_to_subsurface.py
@@ -4,8 +4,8 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 from subsurface.io import WellyToSubsurface
-from subsurface.io.read_to_basic_structs import read_wells_to_unstruct
-from subsurface.io.wells.wells_reader import read_to_welly, read_collar, read_survey
+from subsurface.io.wells import read_wells_to_unstruct
+from subsurface.io.wells.welly_reader import read_to_welly, read_collar, read_survey
 from subsurface.structs import LineSet
 import subsurface
 import pathlib
